home component to render other views. Refactored other views into home/components. Hoisted tests to root view (home). Added navbar tests
