+++ src/views/Home/components/About/index.scss
