+++ src/views/Home/components/FAQs/index.scss
