+++ src/views/Home/components/FAQs/index.view.tsx
