+++ src/views/Home/components/Gallery/index.scss
