+++ src/views/Home/components/Gallery/index.view.tsx
