+++ src/views/Home/components/Projects/index.scss
