+++ src/views/Home/components/Quotes/index.scss
