+++ src/views/Home/components/Quotes/index.view.tsx
