+++ src/views/Home/components/Sponsors/index.scss
