+++ src/views/Home/components/Sponsors/index.view.tsx
