+++ src/views/Team/index.view.tsx
@@ -1,7 +1,7 @@
 import React from "react"
 
 const Team: React.FC = () => {
-  return <h1>Team</h1>
+  return <h1> Our Team</h1>
 }
 
 export default Team
